[REFACTOR] Implement Admin Dashboard with real-time stats, user management, and open complaints tracking

+++ suwmp_be/src/main/java/com/example/suwmp_be/dto/response/DashboardStatsResponse.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DashboardStatsResponse {
     private long totalUsers;
-    private long userGrowth;
+    private Long userGrowth;
     private long activeToday;
-    private int activeTodayGrowth;
+    private Long activeTodayGrowth;
     private long openComplaints;
-    private long openComplaintsDelta;
-    private double systemHealthPercent;
+    private Long openComplaintsDelta;
+    private Double systemHealthPercent;
     private String systemHealthStatus;
 }
